error state correction
