+++ src/components/input/select/style.tsx
@@ -1,25 +1,7 @@
 import styled from 'styled-components';
 import { colors } from '../../colors/colors';
 import { SelectProps } from '.';
-
-const borderStyle = {
-  focused: colors.brand.primaryGradient,
-  error: colors.state.statesError,
-  default: colors.base.grayLight,
-  border: colors.base.grayLight,
-};
-
-const getBorderStyle = (props: SelectProps) => {
-  if (props.focused) {
-    return borderStyle.focused;
-  } else if (props.error) {
-    return borderStyle.error;
-  } else if (props.value) {
-    return borderStyle.default;
-  } else {
-    return borderStyle.border;
-  }
-};
+import { getBorderStyle } from '../../border-state/border-state';
 
 export const StyledSelect = styled.select<SelectProps>`
   width: ${({ expand }) => (expand ? '100%' : 'calc(100% - 25px)')};
