+++ src/components/input/style.tsx
@@ -1,30 +1,12 @@
 import styled from 'styled-components';
 import { InputProps } from '.';
 import { colors } from '../colors/colors';
-
-const borderStyle = {
-  focused: colors.brand.primaryGradient,
-  error: colors.state.statesError,
-  default: colors.base.grayLight,
-  border: colors.base.grayLight,
-};
-
-const getBorderStyle = (props: InputProps) => {
-  if (props.focused) {
-    return borderStyle.focused;
-  } else if (props.error) {
-    return borderStyle.error;
-  } else if (props.value) {
-    return borderStyle.default;
-  } else {
-    return borderStyle.border;
-  }
-};
+import { getBorderStyle } from '../border-state/border-state';
 
 export const StyledInput = styled.input<InputProps>`
   width: ${({ expand }) => (expand ? '100%' : 'calc(100% - 25px)')};
   border-radius: 10px;
-  border: 1px solid ${(props) => getBorderStyle(props)};
+  border: 1px solid ${(props) => getBorderStyle({ ...props, value: props.value ?? '' })};
   background: ${colors.base.background};
   outline: none;
   padding: 12px 20px;
